scripts/spc-load/requirements.txt
locustio==0.14.6
Faker